Updates to filterbanksearch and more work to cleanup old codebase.  Fixes to support filterbank files > 4GB

+++ src/filterbankutil.h
@@ -10,6 +10,7 @@
 #include "filterbank_header.h"
 #include "fitshead.h"
 #include "imswap.h"
+#include <stdint.h>
 
 void filterbank2fits(char * fitsdata, float *datavec, int nchan, int nsamp, long int hitchan, double snr, double doppler, struct filterbank_input *input);
 
@@ -19,6 +20,7 @@ void comp_stats(double *mean, double *stddev, float *vec, long int veclen);
 
 void normalize (float *vec, long int veclen);
 
-long long sizeof_file(char name[]) ;
+long int sizeof_file(char name[]) ;
 
 long int filterbank_extract_from_file(float *output, long int tstart, long int tend, long int chanstart, long int chanend, struct filterbank_input *input);
+
